+++ locales/en/common.json
@@ -13,6 +13,7 @@
   "progress_stepper_current_step": "{{currentStep}} of {{total}} steps completed",
   "required": "Required",
   "return_to_home_page": "Return to home page",
+  "search_button_label": "Search",
   "something_went_wrong": "Something went wrong",
   "submitting_request": "Submitting request",
   "type_value_to_confirm": "Type <i>{{value}}</i> to confirm",
