--- src/ServiceAccount/Search/FilterSelect.tsx
@@ -1,63 +1,0 @@
-import React, { useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import {
-  Select,
-  SelectOption,
-  SelectOptionObject,
-  SelectVariant,
-  ToolbarItem,
-} from '@patternfly/react-core';
-
-export type FilterSelectProps = {
-  filterSelected?: string;
-  setFilterSelected: (value: string) => void;
-};
-
-export const FilterSelect: React.FunctionComponent<FilterSelectProps> = ({
-  setFilterSelected,
-  filterSelected,
-}) => {
-  const { t } = useTranslation();
-  const [isFilterExpanded, setIsFilterExpanded] = useState(false);
-
-  const onToggle = () => {
-    setIsFilterExpanded(!isFilterExpanded);
-  };
-
-  const onSelect = (
-    _event: React.MouseEvent | React.ChangeEvent,
-    selection: string | SelectOptionObject
-  ) => {
-    setIsFilterExpanded(!isFilterExpanded);
-    setFilterSelected(selection?.toString());
-  };
-
-  const options = [
-    { label: t('kafka:short_description'), value: 'description', disabled: false },
-    { label: t('kafka:clientID'), value: 'clientid', disabled: false },
-    { label: t('kafka:owner'), value: 'owner', disabled: false },
-  ];
-
-  return (
-    <ToolbarItem>
-      <Select
-        variant={SelectVariant.single}
-        aria-label='Select filter'
-        onToggle={onToggle}
-        selections={filterSelected}
-        isOpen={isFilterExpanded}
-        onSelect={onSelect}
-      >
-        {options.map((option, index) => (
-          <SelectOption
-            isDisabled={option.disabled}
-            key={index}
-            value={option.value}
-          >
-            {option.label}
-          </SelectOption>
-        ))}
-      </Select>
-    </ToolbarItem>
-  );
-};
