+++ src/ServiceAccount/components/types.ts
@@ -1,0 +1,1 @@
+export type SearchType = "description" | "clientid" | "owner";
